Refactoring Context for modularity
