+++ src/context/rename.rs
@@ -1,0 +1,20 @@
+/*
+ * ANISE Toolkit
+ * Copyright (C) 2021 Christopher Rabotin <christopher.rabotin@gmail.com> et al. (cf. AUTHORS.md)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * Documentation: https://nyxspace.com/
+ */
+
+use crate::asn1::context::AniseContext;
+
+impl<'a> AniseContext<'a> {
+    pub fn rename_ephemeris_traj_mut(&mut self) {
+        todo!()
+    }
+    pub fn rename_orientation_traj_mut(&mut self) {
+        todo!()
+    }
+}
