+++ src/ephemeris.rs
@@ -52,19 +52,19 @@ impl<'a> Ephemeris<'a> {
     }
 
     /// Evaluate this ephemeris at the requested epoch and returns the position only.
-    pub fn position_at(&self, req_epoch: Epoch) -> Result<[f64; 3], AniseError> {
-        let orbit = self.posvel_at(req_epoch)?;
+    pub fn pos(&self, req_epoch: Epoch) -> Result<[f64; 3], AniseError> {
+        let orbit = self.posvel(req_epoch)?;
         Ok([orbit[0], orbit[1], orbit[2]])
     }
 
     /// Evaluate this ephemeris at the requested epoch and returns the velocity only.
-    pub fn velocity_at(&self, req_epoch: Epoch) -> Result<[f64; 3], AniseError> {
-        let orbit = self.posvel_at(req_epoch)?;
+    pub fn vel(&self, req_epoch: Epoch) -> Result<[f64; 3], AniseError> {
+        let orbit = self.posvel(req_epoch)?;
         Ok([orbit[3], orbit[4], orbit[5]])
     }
 
     /// Evaluate this ephemeris at the requested epoch and returns an orbit structure.
-    pub fn posvel_at(&self, req_epoch: Epoch) -> Result<[f64; 6], AniseError> {
+    pub fn posvel(&self, req_epoch: Epoch) -> Result<[f64; 6], AniseError> {
         if self.interpolation_kind != InterpolationKind::ChebyshevSeries {
             return Err(InternalErrorKind::InterpolationNotSupported.into());
         }
